:tada: No more Go! (for now)
After a quick analysis of what this website will have to do, we've decided to revert to a complete JS backend. This website doesn't need speed. With this suite we can deploy this backend with vercel.com for free. We'll use supabase.io as a database, for a free until the 500MB quota.